+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Version 2.0.6 (2017-09-02)
 
+* improve get() behaviour on error with older libvipses
+
 ## Version 2.0.5 (2017-09-02)
 
 * fix get() with older libvipses